+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ### Changed
 
+- Migrate content-grid component to zoneless-ready.
 - Deps: update `@angular/cli` to 20.3.15.
 - Deps: update `htmlparser2` to 10.1.0.
 - Deps (dev): update `@types/node` to 22.19.7.